Change how windows manage the bottom status bar's sprite hiding effect.
Previously enableStatusBarEffect and disableStatusBarEffect would change the line the interrupt would fire at to $8e, which caused the main loop to be perfectly mistimed so it would miss the notification of a completed VBlank. This resulted in two frames worth of processing every other frame. Not a noticible bug until you go looking for cases where you might speed things up by doing more work in a frame.