Fix a bug with status screen and the color port.
A change for the improved status bar included leaving a bad value on the stack. This caused a return to a spot very close to the end of a function working with the Blind status, which didn't cause any negative effects.

Since the Blind effect has been changed heavily for colorization this code changed and caused issues.