Double the number of enemies on the screen where you meet Hasim.
Previously there would be from one to three with even chances of each ammount. Now there will be from two to six with probability of two or six being only 1/9 and probability of four being three times as high.

This is still not meant to be a challenge to players, but is an easy way to eliminate the frequent one-enemy case.