Align metatile data table entries to 16 bytes.
This allows slight performance optimizations, but much more importantly also will allow adding additional data to the entries.

On the other hand, it does take up two and two-thirds the amount of space. If another tileset is added then these will need to be reduced in size somewhat, or split into multiple banks.