Fix an issue with line effects in the color target.
Specifically, letterbox would allow one line of the hidden animation graphics at the top of the screen to show, and the intro scroll effect had single line glitches during the scroll.

What was happening is that the lookup table in the LCDC interrupt was finishing too fast for early entries, leading to the palette values being written a line early. That's been fixed by explicitly waiting for the correct line. Also, a debug message (for color only) has been added in case this interrupt ever ends up late.