Remove script opcode ff from the speedup list.
Further work on issue #23.

Change the speedup list search loop to have a counter rather than relying on the list having a terminator value. The list of opcode numbers to speedup (allow to run if there is extra time in a frame) was terminated with the value ff, which also happens to be a valid script opcode number. Opcode ff is not used directly but is used by other opcodes, including the text control code for printing the Hero or Fuji's name.

Having opcode ff in the speedup list meant that the names were always being printed with no delay between characters.