Remove copying from scriptOpCodeMsg_HandleDualCharacters.
This is part of fixing issue #23.

The original code created a zero terminated buffer for the two characters and passed that to drawText. That meant that a delay when drawText encountered the terminator. The new version takes advantages of changes to drawText and just passes pointers directly to the dual character entry in ROM.