Change printName to not waste time trying to print a null.
Should be the final part of fixing issue #23.

The printName function used a zero terminated buffer and would still call drawText even when the terminator had been reached. It relied on the zero flag from drawText which is set when a terminator is reached or when the end of line is reached, either of which is an acceptable end condition. The downside of that is it caused an extra delay at the end of names.